Initial commit
Built the functioning basics. Next will be plugboard logic.

I think the best way to do plugboard logic is to do a dictionary, but I don't think plugboards had unique letters. Double check that and if they're not unique, then it can be a list of paired values. (ie "TM")

After we handle the swapping behavior of the plugboard, we'll need to simulate the cables and the limits they provided. Check the instructions that were sent with messages; maybe there's a clue into how to represent this on the UI in there.